update Sandbox gate to render optional component when in sandbox mode and add to MainContent
