+++ frontend/components/Sandbox/SandboxDemoMessage/_styles.scss
@@ -1,4 +1,4 @@
-.sandbox-error-message {
+.sandbox-demo-message {
   display: flex;
   flex-direction: column;
   justify-content: center;
