+++ frontend/components/Sandbox/SandboxDemoMessage/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SandboxDemoMessage";
