+++ frontend/pages/admin/UserManagementPage/UserManagementPage.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { InjectedRouter } from "react-router";
-import SandboxGate from "components/SandboxGate";
+import SandboxGate from "components/Sandbox/SandboxGate";
+import SandboxDemoMessage from "components/Sandbox/SandboxDemoMessage";
 import UsersTable from "./components/UsersTable";
 
 const baseClass = "user-management";
@@ -17,8 +18,12 @@ const UserManagementPage = ({ router }: IUserManagementProps): JSX.Element => {
         Fleet.
       </p>
       <SandboxGate
-        message="User management is only available in self-managed Fleet"
-        utmSource="fleet-ui-users-page"
+        fallbackComponent={() => (
+          <SandboxDemoMessage
+            message="User management is only available in self-managed Fleet"
+            utmSource="fleet-ui-users-page"
+          />
+        )}
       >
         <UsersTable router={router} />
       </SandboxGate>
